requirements.txt

selenium
undetected-chromedriver
sqlalchemy
psycopg2-binary
python-dotenv
webdriver-manager
requests
pytz
